Build for all plataforms

+++ OpenML.Shared/Mat3.cpp
@@ -429,7 +429,7 @@ Mat3<T> Mat3<T>::operator-() const
 {
 	Mat3<T> result;
 
-	for (size_t i = 0; i < MAT3_SIZE; i++)
+	for (int i = 0; i < MAT3_SIZE; i++)
 		result[i] = -values[i];
 
 	return result;
@@ -440,7 +440,7 @@ Mat3<T> Mat3<T>::operator-(const Mat3<T>& matrix) const
 {
 	Mat3<T> result;
 
-	for (size_t i = 0; i < MAT3_SIZE; i++)
+	for (int i = 0; i < MAT3_SIZE; i++)
 		result[i] = values[i] - matrix[i];
 
 	return result;
@@ -451,7 +451,7 @@ Mat3<T> Mat3<T>::operator+(const Mat3<T>& matrix) const
 {
 	Mat3<T> result;
 
-	for (size_t i = 0; i < MAT3_SIZE; i++)
+	for (int i = 0; i < MAT3_SIZE; i++)
 		result[i] = values[i] + matrix[i];
 
 	return result;
